requirements.txt
hvac>=1.0.0
boto3>=1.26.0
google-cloud-kms>=2.0.0
cryptography>=3.4.0
